refactor(ops): extract shared workspace::create into claudette::ops
Introduces `claudette::ops` — a process-agnostic command core that the
GUI (Tauri commands), the WS JSON-RPC handler (`claudette-server`), and
the upcoming local-IPC channel for the CLI all dispatch through. Each op
is `(db, hooks, params) -> Result<output>`, with side effects (tray
rebuild, notification sound) delegated through an `OpsHooks` trait so
each surface fires the UI feedback that fits it.

`workspace::create` is the first op extracted. Closes the existing
GUI/server divergence:

- Server's `handle_create_workspace` now returns `default_session_id`
  (it previously returned bare workspace JSON; clients had to follow
  up with `list_chat_sessions` to find the auto-created session id).
- Server now runs the configured setup script (it previously skipped
  setup unconditionally).
- Tauri command continues to fire tray rebuild + SessionStart sound
  via the new `TauriHooks` impl.

Setup script execution is intentionally a separate concern from
`create` itself — callers that have a plugin registry handy (the GUI)
resolve env first, then chain `resolve_and_run_setup`; callers that
don't (the WS server today) pass `None` for env. This avoids forcing
env-provider resolution into the op, which would couple it to
AppState.plugins.

`ops::workspace::create` takes `&mut Database` rather than `&Database`
so the returned future is `Send` — same trick used in
`fork::fork_workspace_at_checkpoint`. `Connection` is `Send` but
`!Sync`, so an exclusive borrow lets it cross awaits.

Branch-prefix helpers (`read_branch_prefix_settings`,
`resolve_branch_prefix`) move from `commands/workspace.rs` into the ops
module so all surfaces resolve prefixes the same way.
